% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{dfMapper_tda}
\alias{dfMapper_tda}
\title{A dataset to test the \code{\link{jaccardMatrix}} and
\code{\link{tdaDfEnrichment}} funtions of \code{PIUMA} package.}
\format{
A data.frame containing 15 rows (cells) and 15 columns (genes)
}
\usage{
dfMapper_tda
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A dataset to test the \code{\link{jaccardMatrix}} and
\code{\link{tdaDfEnrichment}} funtions of \code{PIUMA} package.
}
\keyword{datasets}
