% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{dist_mat_tda}
\alias{dist_mat_tda}
\title{A dataset to test the \code{\link{mapperCore}} funtion of \code{PIUMA}
 package.}
\format{
A data.frame containing 15 rows (cells) and 15 columns (cells)
}
\usage{
dist_mat_tda
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A dataset to test the \code{\link{mapperCore}} funtion of \code{PIUMA}
 package.
}
\keyword{datasets}
