% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getClusters}
\alias{getClusters}
\alias{getClusters,TDAobj-method}
\alias{getClusters,PIUMA-getClusters}
\title{Getter method for the 'clustering' slot of a TDAobj object.}
\usage{
getClusters(x)

\S4method{getClusters}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame
}
\description{
The method to get clusters from the clustering slot
}
\examples{
data(tda_test_data)
}
\author{
Carlo Leonardi
}
