% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getComp}
\alias{getComp}
\alias{getComp,TDAobj-method}
\alias{getComp,PIUMA-getComp}
\title{Getter method for the 'comp' slot of a TDAobj object.}
\usage{
getComp(x)

\S4method{getComp}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the comp data
}
\description{
The method to get data from the comp slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
