% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getDfMapper}
\alias{getDfMapper}
\alias{getDfMapper,TDAobj-method}
\alias{getDfMapper,PIUMA-getDfMapper}
\title{Getter method for the 'dfMapper' slot of a TDAobj object.}
\usage{
getDfMapper(x)

\S4method{getDfMapper}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the dfMapper data
}
\description{
The method to get data from the dfMapper slot
}
\examples{
data(tda_test_data)
ex_out <- getDfMapper(tda_test_data)
}
\author{
Mattia Chiesa
}
