% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getDistMat}
\alias{getDistMat}
\alias{getDistMat,TDAobj-method}
\alias{getDistMat,PIUMA-getDistMat}
\title{Getter method for the 'dist_mat' slot of a TDAobj object.}
\usage{
getDistMat(x)

\S4method{getDistMat}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the dist_mat data
}
\description{
The method to get data from the dist_mat slot
}
\examples{
data(tda_test_data)
ex_out <- getDistMat(tda_test_data)
}
\author{
Mattia Chiesa
}
