% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getGraph}
\alias{getGraph}
\alias{getGraph,TDAobj-method}
\alias{getGraph,PIUMA-getGraph}
\title{Getter method for the 'graph' slot of a TDAobj object.}
\usage{
getGraph(x)

\S4method{getGraph}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
an igraph object
}
\description{
The method to get igraph object from the graph slot
}
\examples{
data(tda_test_data)
}
\author{
Carlo Leonardi
}
