% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getJacc}
\alias{getJacc}
\alias{getJacc,TDAobj-method}
\alias{getJacc,PIUMA-getJacc}
\title{Getter method for the 'jacc' slot of a TDAobj object.}
\usage{
getJacc(x)

\S4method{getJacc}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a matrix with the jacc data
}
\description{
The method to get data from the jacc slot
}
\examples{
data(tda_test_data)
ex_out <- getJacc(tda_test_data)
}
\author{
Mattia Chiesa
}
