% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getMetrics}
\alias{getMetrics}
\alias{getMetrics,TDAobj-method}
\alias{getMetric,PIUMA-getMetrics}
\title{Getter method for the 'metrics' slot under 'graph' of a TDAobj object.}
\usage{
getMetrics(x)

\S4method{getMetrics}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a vector
}
\description{
The method to get metrics from the graph slot
}
\examples{
data(tda_test_data)
}
\author{
Carlo Leonardi, Mattia Chiesa
}
