% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getNodeDataMat}
\alias{getNodeDataMat}
\alias{getNodeDataMat,TDAobj-method}
\alias{getNodeDataMat,PIUMA-getNodeDataMat}
\title{Getter method for the 'node_data_mat' slot of a TDAobj object.}
\usage{
getNodeDataMat(x)

\S4method{getNodeDataMat}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the node_data_mat data
}
\description{
The method to get data from the node_data_mat slot
}
\examples{
data(tda_test_data)
ex_out <- getNodeDataMat(tda_test_data)
}
\author{
Mattia Chiesa
}
