% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getOrigData}
\alias{getOrigData}
\alias{getOrigData,TDAobj-method}
\alias{getOrigData,PIUMA-getOrigData}
\title{Getter method for the 'orig_data' slot of a TDAobj object.}
\usage{
getOrigData(x)

\S4method{getOrigData}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the original data
}
\description{
The method to get data from the orig_data slot
}
\examples{
data(tda_test_data)
ex_out <- getOrigData(tda_test_data)
}
\author{
Mattia Chiesa
}
