% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getOutcome}
\alias{getOutcome}
\alias{getOutcome,TDAobj-method}
\alias{getOutcome,PIUMA-getOutcome}
\title{Getter method for the 'outcome' slot of a TDAobj object.}
\usage{
getOutcome(x)

\S4method{getOutcome}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the outcome data
}
\description{
The method to get data from the outcome slot
}
\examples{
data(tda_test_data)
ex_out <- getOutcome(tda_test_data)
}
\author{
Mattia Chiesa
}
