% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getOutcomeFact}
\alias{getOutcomeFact}
\alias{getOutcomeFact,TDAobj-method}
\alias{getOutcomeFact,PIUMA-getOutcomeFact}
\title{Getter method for the 'outcomeFact' slot of a TDAobj object.}
\usage{
getOutcomeFact(x)

\S4method{getOutcomeFact}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the outcomeFact data
}
\description{
The method to get data from the outcomeFact slot
}
\examples{
data(tda_test_data)
ex_out <- getOutcomeFact(tda_test_data)
}
\author{
Mattia Chiesa
}
