% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{getScaledData}
\alias{getScaledData}
\alias{getScaledData,TDAobj-method}
\alias{getScaledData,PIUMA-getScaledData}
\title{Getter method for the 'scaled_data' slot of a TDAobj object.}
\usage{
getScaledData(x)

\S4method{getScaledData}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a data.frame with the scaled data
}
\description{
The method to get data from the scaled_data slot
}
\examples{
data(tda_test_data)
ex_out <- getScaledData(tda_test_data)
}
\author{
Mattia Chiesa
}
