% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Import.R
\name{importSplitScale}
\alias{importSplitScale}
\title{Import and handle dataframes for TDA Mapper}
\usage{
importSplitScale(df, outcomes)
}
\arguments{
\item{df}{A data.frame representing a dataset in the classical n x m form.
Rows (n) and columns (m) should be, respectively, observations and features.}

\item{outcomes}{A string or vector of string containing the name of variables
that have to be considered 'outcomes'}
}
\value{
A list containing:
\itemize{
  \item A data.frame of original data (without outcomes).
  \item A data.frame of standardized data (without outcomes)
  \item A data.frame of re-scaled data (without outcomes)
  \item A data.frame of original outcomes.
  \item A data.frame of original outcomes converted as numeric.
}
}
\description{
This function import a data.frame and performs some preliminary
steps on it. specifically, outcomes variables data will be separated by
the rest of data. The remaining dataset will be also re-scaled (0-1) and
standardized (z-score).
}
\examples{
## use example data:
data("vascEC_meta")
data("vascEC_norm")
df <- cbind(vascEC_meta,vascEC_norm)
res <- importSplitScale(df, "zone")

}
\author{
Laura Ballarini, Mattia Chiesa
}
