% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{jacc_mat_tda}
\alias{jacc_mat_tda}
\title{A dataset to test the \code{\link{checkScaleFreeModel}} funtion
of \code{PIUMA} package.}
\format{
A data.frame containing 17 rows (nodes) and 17 columns (nodes)
}
\usage{
jacc_mat_tda
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A dataset to test the \code{\link{checkScaleFreeModel}} funtion
of \code{PIUMA} package.
}
\keyword{datasets}
