% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{lenses_tda}
\alias{lenses_tda}
\title{A dataset to test the \code{\link{mapperCore}} funtion of \code{PIUMA}
 package.}
\format{
A data.frame containing 15 rows (cells) and 2 columns (principal
components)
}
\usage{
lenses_tda
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A dataset to test the \code{\link{mapperCore}} funtion of \code{PIUMA}
 package.
}
\keyword{datasets}
