% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_namespace.R
\docType{package}
\name{PIUMA-package}
\alias{PIUMA-package}
\title{PIUMA: Phenotypes Identification Using Mapper from topological data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The PIUMA package offers a tidy pipeline of Topological Data Analysis frameworks to identify and characterize communities in high and heterogeneous dimensional data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioinfoMonzino/PIUMA}
  \item Report bugs at \url{https://github.com/BioinfoMonzino/PIUMA/issues}
}

}
\author{
\strong{Maintainer}: Mattia Chiesa \email{mattia.chiesa@cardiologicomonzino.it} (\href{https://orcid.org/0000-0001-7427-9954}{ORCID})

Authors:
\itemize{
  \item Arianna Dagliati \email{arianna.dagliati@unipv.it} (\href{https://orcid.org/0000-0002-5041-0409}{ORCID})
  \item Alessia Gerbasi \email{alessia.gerbasi01@universitadipavia.it} (\href{https://orcid.org/0000-0003-4501-1777}{ORCID})
  \item Giuseppe Albi \email{giuseppe.albi01@universitadipavia.it}
  \item Laura Ballarini \email{laura.ballarini01@universitadipavia.it}
  \item Luca Piacentini \email{luca.piacentini@cardiologicomonzino.it} (\href{https://orcid.org/0000-0003-1022-4481}{ORCID})
  \item Carlo Leonardi \email{cl35@sanger.ac.uk} (\href{https://orcid.org/0000-0001-5348-8300}{ORCID})
}

}
\keyword{internal}
