% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_packagepage.R
\docType{package}
\name{PIUMA}
\alias{PIUMA}
\title{PIUMA: Phenotypes Identification Using Mapper from topological data Analysis}
\description{
The application of unsupervised learning methodologies could help the
identification of specific phenotypes in huge heterogeneous cohorts,
such as clinical or -omics data. Among them, the Topological Data Analysis
 (TDA) is a rapidly growing field that combines concepts from algebraic
topology and computational geometry to analyze and extract meaningful
information from complex and high-dimensional data sets. Moreover,
TDA is a robust and effective methodology, able to preserve the
intrinsic characteristics of data and the mutual relations among
 observations, depicting complex data in a graph-based representation.
 Indeed, building topological models as networks, TDA allows complex
 diseases to be inspected in a continuous space, where subjects can
 fluctuate over the graph, sharing, at the same time, more than one
 adjacent node of the network. Overall, TDA offers a powerful set
 of tools to capture the underlying topological features of data,
 revealing essential patterns and relationships that might be
 hidden from traditional statistical techniques.
 The PIUMA package (Phenotypes Identification Using Mapper from
 topological data Analysis) allows implementing all the main
 steps of a Topological Data Analysis.
 PIUMA is the italian word meaning 'feather'.
}
\details{
See the package vignette, by typing \code{vignette("PIUMA")} to discover
all the functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioinfoMonzino/PIUMA}
  \item Report bugs at \url{https://github.com/BioinfoMonzino/PIUMA/issues}
}

}
\author{
Mattia Chiesa, Laura Ballarini, Luca Piacentini
}
\keyword{package}
