% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setComp}
\alias{setComp}
\alias{setComp,TDAobj-method}
\alias{setComp,PIUMA-setComp}
\title{Setter method for the 'comp' slot of a TDAobj object.}
\usage{
setComp(x, y)

\S4method{setComp}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the comp data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the comp slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
