% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setDfMapper}
\alias{setDfMapper}
\alias{setDfMapper,TDAobj-method}
\alias{setDfMapper,PIUMA-setDfMapper}
\title{Setter method for the 'dfMapper' slot of a TDAobj object.}
\usage{
setDfMapper(x, y)

\S4method{setDfMapper}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the dfMapper data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the dfMapper slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
