% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setDistMat}
\alias{setDistMat}
\alias{setDistMat,TDAobj-method}
\alias{setDistMat,PIUMA-setDistMat}
\title{Setter method for the 'dist_mat' slot of a TDAobj object.}
\usage{
setDistMat(x, y)

\S4method{setDistMat}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the dist_mat data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the dist_mat slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
