% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setGraph}
\alias{setGraph}
\alias{setGraph,TDAobj-method}
\alias{setGraph,PIUMA-setGraph}
\title{Setter method for the 'graph' slot of a TDAobj object.}
\usage{
setGraph(x)

\S4method{setGraph}{TDAobj}(x)
}
\arguments{
\item{x}{a \code{TDAobj} object}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set igraph object to the graph slot
}
\examples{
data(tda_test_data)
}
\author{
Carlo Leonardi
}
