% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setJacc}
\alias{setJacc}
\alias{setJacc,TDAobj-method}
\alias{setJacc,PIUMA-setJacc}
\title{Setter method for the 'jacc' slot of a TDAobj object.}
\usage{
setJacc(x, y)

\S4method{setJacc}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a matrix with the jacc data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the jacc slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
