% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setNodeDataMat}
\alias{setNodeDataMat}
\alias{setNodeDataMat,TDAobj-method}
\alias{setNodeDataMat,PIUMA-setNodeDataMat}
\title{Setter method for the 'node_data_mat' slot of a TDAobj object.}
\usage{
setNodeDataMat(x, y)

\S4method{setNodeDataMat}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the node_data_mat data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the node_data_mat slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
