% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setOrigData}
\alias{setOrigData}
\alias{setOrigData,TDAobj-method}
\alias{setOrigData,PIUMA-setOrigData}
\title{Setter method for the 'orig_data' slot of a TDAobj object.}
\usage{
setOrigData(x, y)

\S4method{setOrigData}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the original data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the orig_data slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
