% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setOutcome}
\alias{setOutcome}
\alias{setOutcome,TDAobj-method}
\alias{setOutcome,PIUMA-setOutcome}
\title{Setter method for the 'outcome' slot of a TDAobj object.}
\usage{
setOutcome(x, y)

\S4method{setOutcome}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the outcome data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the outcome slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
