% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setOutcomeFact}
\alias{setOutcomeFact}
\alias{setOutcomeFact,TDAobj-method}
\alias{setOutcomeFact,PIUMA-setOutcomeFact}
\title{Setter method for the 'outcomeFact' slot of a TDAobj object.}
\usage{
setOutcomeFact(x, y)

\S4method{setOutcomeFact}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the outcomeFact data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the outcomeFact slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
