% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Generics.R, R/PIUMA_Methods.R
\docType{methods}
\name{setScaledData}
\alias{setScaledData}
\alias{setScaledData,TDAobj-method}
\alias{setScaledData,PIUMA-setScaledData}
\title{Setter method for the 'scaled_data' slot of a TDAobj object.}
\usage{
setScaledData(x, y)

\S4method{setScaledData}{TDAobj}(x, y)
}
\arguments{
\item{x}{a \code{TDAobj} object}

\item{y}{a data.frame with the scaled data}
}
\value{
a \code{TDAobj} object
}
\description{
The method to set the scaled_data slot
}
\examples{
data(tda_test_data)
}
\author{
Mattia Chiesa
}
