% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{tda_test_data}
\alias{tda_test_data}
\title{A TDAobj to test the \code{PIUMA} package.}
\format{
A \code{TDAobj}.
}
\usage{
data(tda_test_data)
}
\description{
A \code{TDAobj} with data in all slots for testing.
}
\keyword{datasets}
