% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{vascEC_norm}
\alias{vascEC_norm}
\title{We tested PIUMA on a subset of the single-cell RNA Sequencing dataset
(GSE:GSE193346 generated and published by Feng et al. (2022)
on Nature Communication to demonstrate that distinct
transcriptional profiles are present in specific cell types of each heart
chambers, which were attributed to have roles in cardiac development.
In this tutorial, our aim will be to exploit PIUMA for identifying
sub-population of vascular endothelial cells, which can be associated
with specific heart developmental stages. The original dataset consisted
of three layers of heterogeneity: cell type, stage and zone
(i.e., heart chamber). Our testing dataset was obtained by
subsetting vascular endothelial cells (cell type) by Seurat object,
extracting raw counts and metadata. Thus, we filtered low expressed genes
and normalized data by DaMiRseq}
\format{
A matrix with 1180 rows (cells) and 838 columns (genes).
}
\usage{
data(vascEC_norm)
}
\description{
We tested PIUMA on a subset of the single-cell RNA Sequencing dataset
(GSE:GSE193346 generated and published by Feng et al. (2022)
on Nature Communication to demonstrate that distinct
transcriptional profiles are present in specific cell types of each heart
chambers, which were attributed to have roles in cardiac development.
In this tutorial, our aim will be to exploit PIUMA for identifying
sub-population of vascular endothelial cells, which can be associated
with specific heart developmental stages. The original dataset consisted
of three layers of heterogeneity: cell type, stage and zone
(i.e., heart chamber). Our testing dataset was obtained by
subsetting vascular endothelial cells (cell type) by Seurat object,
extracting raw counts and metadata. Thus, we filtered low expressed genes
and normalized data by DaMiRseq
}
\keyword{datasets}
