% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_assessTDA.R
\name{checkNetEntropy}
\alias{checkNetEntropy}
\title{Compute the Network Entropy (only for binary outcome!)}
\usage{
checkNetEntropy(vectFreq)
}
\arguments{
\item{vectFreq}{A vector containing the frequency of the outcome
(i.e. p(outcome=1) ) for each node of a network.}
}
\value{
The average of the entropies computed for each node of a network.
}
\description{
This function computes the average of the entropies computed for
each node of a network, only for variables whose elements range (0;1)
}
\details{
The average of the entropies is related to the amount of information
stored in the network.
}
\examples{
# use example data:
set.seed(1)
entropy <- checkNetEntropy(runif(5))

}
\seealso{
\code{\link{importSplitScale}},
\code{\link{dfToDistance}},
\code{\link{dfToProjection}},
\code{\link{mapperCore}},
\code{\link{jaccardMatrix}},
\code{\link{tdaDfEnrichment}}
}
\author{
Laura Ballarini, Mattia Chiesa
}
