% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{df_test_proj}
\alias{df_test_proj}
\title{A dataset to test the \code{\link{dfToProjection}} and
\code{\link{dfToDistance}} funtions of \code{PIUMA} package.}
\format{
A data.frame containing 15 rows (cells) and 15 columns (genes)
}
\usage{
df_test_proj
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A dataset to test the \code{\link{dfToProjection}} and
\code{\link{dfToDistance}} funtions of \code{PIUMA} package.
}
\keyword{datasets}
