% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Distances.R
\name{dfToDistance}
\alias{dfToDistance}
\title{Compute the Distance Matrix from a dataset}
\usage{
dfToDistance(df, distMethod = c("euclidean", "gower", "pearson"))
}
\arguments{
\item{df}{A data.frame representing a dataset in the classical n x m form.
Rows (n) and columns (m) should be, respectively, observations and features.}

\item{distMethod}{The distance method to calculate the distance matix.
"euclidean", "gower" and "pearson" values are allowed. Default: "euclidean".}
}
\value{
A data.frame containing the distance matrix computed
}
\description{
This function returns the distance matrix computed by using the
 Pearson's, Euclidean or Gower distance methods. The distances are
computed between the rows of a data.frame in the classical form n x m,
 where n (rows) are observations and m (columns) are features.
}
\examples{
## use example data:
data(df_test_proj)
dfDist <- dfToDistance(df_test_proj, "euclidean")

}
\seealso{
\code{\link{importSplitScale}}
}
\author{
Laura Ballarini, Mattia Chiesa
}
