% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{tda_test_data}
\alias{tda_test_data}
\title{A TDAobj to test the \code{PIUMA} package.}
\format{
A TDAobj with data in all slots
}
\usage{
tda_test_data
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A TDAobj to test the \code{PIUMA} package.
}
\keyword{datasets}
