% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_Enrichment.R
\name{tdaDfEnrichment}
\alias{tdaDfEnrichment}
\title{Add information to TDA dataframe}
\usage{
tdaDfEnrichment(dfMapper, df)
}
\arguments{
\item{dfMapper}{A data.frame generated by the \code{\link{mapperCore}}
function.}

\item{df}{A data.frame with scaled vslues (\code{\link{importSplitScale}})
in the classical n) x m form: rows (n) and columns (m) must be observations 
and features, respectively.}
}
\value{
A data.frame with the average value and the size.
}
\description{
This function computes the average value of the features
and the size of each rows of the dataframe 
generated by the \code{\link{mapperCore}} function.
}
\examples{
## use example data:
data(dfMapper_tda)
data(df_test_proj)
enrich_mat_tda <- tdaDfEnrichment(dfMapper_tda, df_test_proj)

}
\seealso{
\code{\link{importSplitScale}},
\code{\link{dfToDistance}},
\code{\link{dfToProjection}},
\code{\link{mapperCore}},
\code{\link{jaccardMatrix}}
}
\author{
Laura Ballarini, Mattia Chiesa
}
