% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_TDAobj.R
\docType{class}
\name{TDAobj-class}
\alias{TDAobj-class}
\title{The object 'TDAobj'}
\value{
TDAobj class
showClass("TDAobj")
}
\description{
The TDA object for storing TDA data
}
\section{Slots}{

\describe{
\item{\code{orig_data}}{A data.frame of original data (without outcomes)}

\item{\code{scaled_data}}{A data.frame of re-scaled data (without outcomes)}

\item{\code{outcomeFact}}{A data.frame of original outcomes}

\item{\code{outcome}}{A data.frame of original outcomes converted as numeric}

\item{\code{comp}}{A data.frame containing the components of projected data}

\item{\code{dist_mat}}{A data.frame containing the computed distance matrix}

\item{\code{dfMapper}}{A data.frame containing the nodes, with their elements,
identified by TDA}

\item{\code{jacc}}{A matrix of Jaccard indexes between each pair of dfMapper nodes}

\item{\code{node_data_mat}}{A data.frame with the node size and the average value
of each feature}
}}

