% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{vascEC_meta}
\alias{vascEC_meta}
\title{Example datasets for PIUMA package}
\format{
A dataframe containing 1180 rows (cells) and 2 columns (outcomes)
}
\usage{
vascEC_meta
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A dataset with metadata extracted from Feng et al.
(2022, Nature Communications) to test the
\code{\link{importSplitScale}} funtion of \code{PIUMA} package.
}
\keyword{datasets}
