% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUMA_testdata.R
\docType{data}
\name{vascEC_norm}
\alias{vascEC_norm}
\title{A matrix with scRNA-seq counts data extracted from Feng et al.
(2022, Nature Communications) to test the
\code{\link{importSplitScale}} funtion of \code{PIUMA} package.}
\format{
A matrix containing 1180 rows (cells) and 838 columns (genes)
}
\usage{
vascEC_norm
}
\value{
An example dataset for \code{PIUMA} package
}
\description{
A matrix with scRNA-seq counts data extracted from Feng et al.
(2022, Nature Communications) to test the
\code{\link{importSplitScale}} funtion of \code{PIUMA} package.
}
\keyword{datasets}
